+++ CHANGELOG.md
@@ -21,12 +21,15 @@
 ### Fixed
 
 - [TD-4721] User with permissions can list sources
+
+# Add
+
 - [TD-4695] Structure bulk update was not reindexing structures with updated
   notes
 - [TD-3128] Profile `null_count`, `total_count` and `unique_count` fields were
   limited to PostgreSQL `integer` values (maximum 2,147,483,647). They have now
   been changed to `bigint` (maximum 9,223,372,036,854,775,807)
-
+- [TD-4538] Segments definitions in the implementations, and support for segment results
 ## [4.41.1] 2022-04-04
 
 # Added